RaspberryPi/Init/requirementsPi.txt
gpiozero
threaded
sockets
slack
regex
paho-mqtt
typing_extensions
requests
flask
pytapo
slackclient
pyOpenSSL
pygame
mfrc522